pyproject.toml
[project]
name = "anonymize-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clevercsv>=0.8.3",
    "loguru>=0.7.3",
    "matplotlib>=3.10.6",
    "opencv-contrib-python-headless>=4.12.0.88",
    "paddleocr>=3.2.0",
    "paddlepaddle==3.0.0",
    "presidio-analyzer>=2.2.360",
    "presidio-anonymizer>=2.2.360",
    "presidio-image-redactor>=0.0.57",
    "pydicom>=3.0.1",
    "typer-slim>=0.19.2",
]

[dependency-groups]
dev = [
    "ipython>=9.6.0",
]
